backend/requirements.txt
fastapi
uvicorn
python-multipart
gpt4all
python-dotenv
